ANDKR-1167 | 위도 경도 double로 변경

+++ dailyHOTEL/app/src/main/java/com/daily/dailyhotel/repository/remote/model/GourmetSalesData.java
@@ -67,10 +67,10 @@ public class GourmetSalesData
     public String lastOrderTime;
 
     @JsonField(name = "latitude")
-    public int latitude;
+    public double latitude;
 
     @JsonField(name = "longitude")
-    public int longitude;
+    public double longitude;
 
     @JsonField(name = "menuBenefit")
     public String menuBenefit;
